// apps/packages/app/src/api/files/v2/external/type.ts
import { DriveType } from 'src/utils/interface/files';
export interface DriveMenuType {
	label: string;
	key: string | number;
	icon: string;
	driveType: DriveType;
}

export interface SmbMountType {
	password?: string;
	timestamp?: number;
	url: string;
	username?: string;
}

export const normalExternalTypes = ['internal'];